+++ native/CHANGELOG.md
@@ -6,7 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## Recent Changes
 
-- `c`: Fixed an issue where `zoweax` could run the `server` command or `plugin` command group as APF-authorized. Now, `zoweax` disables APF authorization before commands are executed. Commands in the `console` group continue to run as APF authorized. [#958](https://github.com/zowe/zowex/issues/958)
+- `c`: Fixed an issue where `zoweax` could run the `server` command or `plugin` command group as privileged. Now, `zoweax` disables privileges before commands are executed. Commands in the `console` group continue to run as privileged. [#958](https://github.com/zowe/zowex/issues/958)
 - `c`: Added support for pre-command hooks in the command-line parser utility. [#958](https://github.com/zowe/zowex/issues/958)
 
 ## `0.5.0`